changes in bookings

+++ bookings/views.py
@@ -17,7 +17,7 @@ from bookings.utils import get_booking_by_id
 from datetime import datetime
 
 @secure_required
-@permission_required_or_403('insite_messages.view_message')
+@permission_required_or_403('bookings.view_booking')
 def bookings(request, username,
                 template_name='bookings/bookings.html',
                 extra_context=None, **kwargs):
@@ -38,6 +38,7 @@ def bookings(request, username,
                                           extra_context=extra_context)(request)
 
 @secure_required
+@permission_required_or_403('bookings.add_booking')
 def booking_add(request, username, booking_form = BookingForm,
                 template_name = 'bookings/booking_add.html', success_url = None,
                 extra_context=None, **kwargs):
@@ -59,6 +60,7 @@ def booking_add(request, username, booking_form = BookingForm,
             redirect_to = success_url
           else:
             redirect_to = reverse('profiles:bookings', kwargs={'username': username})
+
           return redirect(redirect_to)
 
 
@@ -72,19 +74,39 @@ def booking_add(request, username, booking_form = BookingForm,
 
     return ExtraContextTemplateView.as_view(template_name=template_name,
                                           extra_context=extra_context)(request)
-@secure_required
-def booking_add_with_service(request, username, service_id):
-    pass
 
 @secure_required
+@permission_required_or_403('bookings.reject_booking')
 def booking_reject(request, username, booking_id):
-    pass
+
+    user = get_object_or_404(User, username__iexact=username)
+    booking = get_booking_by_id(booking_id)
+
+    if user:
+        if user.has_perm('reject_booking', booking):
+            booking.status = 4
+            booking.save()
+
+    url = reverse('profiles:bookings', kwargs={'username':user.username})
+    return HttpResponseRedirect(url)
 
 @secure_required
+@permission_required_or_403('bookings.approve_booking')
 def booking_approve(request, username, booking_id):
-    pass
+
+    user = get_object_or_404(User, username__iexact=username)
+    booking = get_booking_by_id(booking_id)
+
+    if user:
+        if user.has_perm('approve_booking', booking):
+            booking.status = 3
+            booking.save()
+
+    url = reverse('profiles:bookings', kwargs={'username':user.username})
+    return HttpResponseRedirect(url)
 
 @secure_required
+@permission_required_or_403('bookings.view_booking')
 def booking_view(request, username, booking_id, template_name = 'bookings/booking_view.html', success_url=None,
                  extra_context=None, **kwargs):
 
@@ -101,7 +123,6 @@ def booking_view(request, username, booking_id, template_name = 'bookings/bookin
 
     # if user.has_perm('bookings.view_message', booking):
     extra_context['booking'] = booking
-
     extra_context['profile'] = profile
 
     extra_context = makeContextForDetails(request, extra_context)
@@ -111,6 +132,7 @@ def booking_view(request, username, booking_id, template_name = 'bookings/bookin
                                           extra_context=extra_context)(request)
 
 @secure_required
+@permission_required_or_403('bookings.change_booking')
 def booking_edit(request, username, booking_id, edit_booking_form = BookingForm,
                  template_name ='bookings/booking_edit.html', success_url = None,
                   extra_context=None, **kwargs):
@@ -118,13 +140,16 @@ def booking_edit(request, username, booking_id, edit_booking_form = BookingForm,
     user = get_object_or_404(User, username__iexact=username)
     profile = get_user_profile(user)
     booking = get_booking_by_id(booking_id)
+
     form = edit_booking_form(instance=booking)
 
     if request.method == 'POST':
         form = edit_booking_form(request.POST, request.FILES, instance=booking)
 
         if form.is_valid():
-            booking = form.save(user)
+            if user.has_perm('change_booking', booking):
+                if(booking.status == 1 or booking.status == 2):
+                    booking = form.save(user, booking)
 
             if success_url:
                 redirect_to = success_url
@@ -145,17 +170,19 @@ def booking_edit(request, username, booking_id, edit_booking_form = BookingForm,
 
 
 @secure_required
-def booking_remove(request,booking_id, template_name='bookings/bookings.html', success_url=None,
+@permission_required_or_403('bookings.delete_booking')
+def booking_remove(request, username, booking_id, template_name='bookings/bookings.html', success_url=None,
                  extra_context=None, **kwargs):
 
-    user = request.user
-    booking = Booking.get(pk = booking_id)
+    user = get_object_or_404(User, username__iexact=username)
+    booking = Booking.objects.get(pk = booking_id)
 
     if user:
         if user.has_perm('delete_booking', booking):
-            booking.delete()
+            if booking.sender == user:
+                booking.delete()
 
-    url = reverse('bookings:bookings', kwargs={'username':user.username})
+    url = reverse('profiles:bookings', kwargs={'username':user.username})
     return HttpResponseRedirect(url)
